add classifications and getfiles  AB#207 AB#208

+++ src/Services/Lexon/Lexon.MySql/Model/JosFile.cs
@@ -6,6 +6,6 @@ namespace Lexon.MySql.Model
         public string code { get; set; }
         public string mails { get; set; }
         public string Description { get; set; }
-        public long idFile { get; set; }
+        public int idFile { get; set; }
     }
 }
